# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Autobroker API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /lots:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: q, schema: { type: string } }
        - { in: query, name: status_order, schema: { type: string } }
        - { in: query, name: archived, schema: { type: boolean } }
      responses: { "200": { description: paginated lots } }
    post:
      security: [{ bearerAuth: [] }]
      responses: { "201": { description: created } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer